+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Provide helper method on FormElementRenderer to provide the criteria-based highlight classes
 * Include highlight_if and hide_display_if in the field element models
   (we need access to them for display-mode code)
 * Refactor config to allow easy use and overriding of defaults